revno: 1268 message:   Improved performance in outlier analysis
committer: Lars Helge Oeverland <larshelge@gmail.com>
branch nick: trunk
timestamp: Wed 2009-12-23 13:09:02 +0100

+++ dhis-2/dhis-api/src/main/java/org/hisp/dhis/outlieranalysis/OutlierAnalysisStore.java
@@ -45,5 +45,5 @@ public interface OutlierAnalysisStore
     Double getAverage( DataElement dataElement, DataElementCategoryOptionCombo categoryOptionCombo, OrganisationUnit organisationUnit );
     
     Collection<DeflatedDataValue> getDeflatedDataValues( DataElement dataElement, DataElementCategoryOptionCombo categoryOptionCombo,
-        Collection<Period> periods, OrganisationUnit organisationUnit, double lowerBound, double upperBound );
+        Collection<Period> periods, OrganisationUnit organisationUnit, int lowerBound, int upperBound );
 }
